Adapting test to the package refactor
